+++ tests/tools.go
@@ -1,4 +1,4 @@
-package main
+package tests
 
 import (
     "fmt"
